pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-review-graph"
version = "1.8.4"
description = "Persistent incremental knowledge graph for token-efficient, context-aware code reviews with Claude Code"
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Tirth" },
]
keywords = ["code-review", "knowledge-graph", "tree-sitter", "claude-code", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "mcp>=1.0.0,<2",
    "fastmcp>=0.1.0,<2",
    "tree-sitter>=0.23.0,<1",
    "tree-sitter-language-pack>=0.3.0,<1",
    "networkx>=3.2,<4",
    "watchdog>=4.0.0,<6",
]

[project.urls]
Homepage = "https://github.com/tirth8205/code-review-graph"
Repository = "https://github.com/tirth8205/code-review-graph"
Documentation = "https://github.com/tirth8205/code-review-graph/blob/main/docs/INDEX.md"
Changelog = "https://github.com/tirth8205/code-review-graph/blob/main/CHANGELOG.md"
Issues = "https://github.com/tirth8205/code-review-graph/issues"

[project.scripts]
code-review-graph = "code_review_graph.cli:main"

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0.0,<4",
    "numpy>=1.26,<3",
]
google-embeddings = [
    "google-generativeai>=0.8.0,<1",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.3.0,<1",
]
langgraph-agent = [
    "langgraph>=0.2,<1",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.2,<0.4",
    "typing_extensions>=4.8,<5",
    "streamlit>=1.28,<2",
    "python-dotenv>=1.0,<2",
]

[tool.hatch.build.targets.wheel]
packages = ["code_review_graph"]

[tool.hatch.build.targets.sdist]
include = [
    "code_review_graph/",
    "skills/",
    ".claude-plugin/",
    "docs/",
    "hooks/",
    "LICENSE",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"code_review_graph/visualization.py" = ["E501"]  # embedded HTML/JS template

[tool.bandit]
# B101: assert used (fine in non-security code)
# B404: import subprocess (we need git interaction)
# B603: subprocess without shell=True (we use list args, not shell)
# B607: partial executable path (calling "git" by name is standard)
# B608: SQL f-string — false positive, we use parameterized "?" placeholders on a local SQLite DB
skips = ["B101", "B404", "B603", "B607", "B608"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["tests/fixtures"]
